+++ bundles/itemsbundle/items_dbhelper.go
@@ -3,7 +3,7 @@ package itemsbundle
 import (
 	"log"
 
-	"github.com/osoderholm/eletab-lite/app/common"
+	"github.com/osoderholm/eletab-lite/common"
 )
 
 // Empty helper struct for DBHelper interface
