+++ common/db_helper.go
